# internal/adapters/storage/postgres/generated/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package generated

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type AttendanceLog struct {
	ID              pgtype.UUID      `json:"id"`
	StudentID       pgtype.UUID      `json:"student_id"`
	CameraID        pgtype.UUID      `json:"camera_id"`
	CheckInTime     pgtype.Timestamp `json:"check_in_time"`
	ConfidenceScore pgtype.Float8    `json:"confidence_score"`
	SnapshotUrl     pgtype.Text      `json:"snapshot_url"`
	IsValid         pgtype.Bool      `json:"is_valid"`
	CreatedAt       pgtype.Timestamp `json:"created_at"`
}

type AttendanceLogs202401 struct {
	ID              pgtype.UUID      `json:"id"`
	StudentID       pgtype.UUID      `json:"student_id"`
	CameraID        pgtype.UUID      `json:"camera_id"`
	CheckInTime     pgtype.Timestamp `json:"check_in_time"`
	ConfidenceScore pgtype.Float8    `json:"confidence_score"`
	SnapshotUrl     pgtype.Text      `json:"snapshot_url"`
	IsValid         pgtype.Bool      `json:"is_valid"`
	CreatedAt       pgtype.Timestamp `json:"created_at"`
}

type AttendanceLogs202402 struct {
	ID              pgtype.UUID      `json:"id"`
	StudentID       pgtype.UUID      `json:"student_id"`
	CameraID        pgtype.UUID      `json:"camera_id"`
	CheckInTime     pgtype.Timestamp `json:"check_in_time"`
	ConfidenceScore pgtype.Float8    `json:"confidence_score"`
	SnapshotUrl     pgtype.Text      `json:"snapshot_url"`
	IsValid         pgtype.Bool      `json:"is_valid"`
	CreatedAt       pgtype.Timestamp `json:"created_at"`
}

type AttendanceLogsDefault struct {
	ID              pgtype.UUID      `json:"id"`
	StudentID       pgtype.UUID      `json:"student_id"`
	CameraID        pgtype.UUID      `json:"camera_id"`
	CheckInTime     pgtype.Timestamp `json:"check_in_time"`
	ConfidenceScore pgtype.Float8    `json:"confidence_score"`
	SnapshotUrl     pgtype.Text      `json:"snapshot_url"`
	IsValid         pgtype.Bool      `json:"is_valid"`
	CreatedAt       pgtype.Timestamp `json:"created_at"`
}

type AuditLog struct {
	ID        int64            `json:"id"`
	UserID    pgtype.UUID      `json:"user_id"`
	Action    string           `json:"action"`
	TableName pgtype.Text      `json:"table_name"`
	RecordID  pgtype.Text      `json:"record_id"`
	OldValue  []byte           `json:"old_value"`
	NewValue  []byte           `json:"new_value"`
	IpAddress pgtype.Text      `json:"ip_address"`
	UserAgent pgtype.Text      `json:"user_agent"`
	CreatedAt pgtype.Timestamp `json:"created_at"`
}

type Camera struct {
	ID        pgtype.UUID      `json:"id"`
	ZoneID    pgtype.UUID      `json:"zone_id"`
	Name      string           `json:"name"`
	IpAddress pgtype.Text      `json:"ip_address"`
	RtspUrl   string           `json:"rtsp_url"`
	Status    pgtype.Text      `json:"status"`
	AiEnabled pgtype.Bool      `json:"ai_enabled"`
	CreatedAt pgtype.Timestamp `json:"created_at"`
	UpdatedAt pgtype.Timestamp `json:"updated_at"`
}

type Class struct {
	ID                pgtype.UUID      `json:"id"`
	Name              string           `json:"name"`
	Grade             int32            `json:"grade"`
	AcademicYear      pgtype.Text      `json:"academic_year"`
	HomeroomTeacherID pgtype.UUID      `json:"homeroom_teacher_id"`
	CreatedAt         pgtype.Timestamp `json:"created_at"`
}

type DailyAttendance struct {
	ID           int32            `json:"id"`
	StudentID    pgtype.UUID      `json:"student_id"`
	ClassID      pgtype.UUID      `json:"class_id"`
	Date         pgtype.Date      `json:"date"`
	FirstCheckIn pgtype.Timestamp `json:"first_check_in"`
	LastCheckOut pgtype.Timestamp `json:"last_check_out"`
	Status       pgtype.Text      `json:"status"`
	Note         pgtype.Text      `json:"note"`
	CreatedAt    pgtype.Timestamp `json:"created_at"`
	UpdatedAt    pgtype.Timestamp `json:"updated_at"`
}

type Role struct {
	ID          pgtype.UUID      `json:"id"`
	Name        string           `json:"name"`
	Description pgtype.Text      `json:"description"`
	CreatedAt   pgtype.Timestamp `json:"created_at"`
}

type Student struct {
	ID             pgtype.UUID      `json:"id"`
	StudentCode    string           `json:"student_code"`
	FullName       string           `json:"full_name"`
	ClassID        pgtype.UUID      `json:"class_id"`
	DateOfBirth    pgtype.Date      `json:"date_of_birth"`
	Gender         pgtype.Text      `json:"gender"`
	AvatarUrl      pgtype.Text      `json:"avatar_url"`
	ParentPhone    pgtype.Text      `json:"parent_phone"`
	FaceRegistered pgtype.Bool      `json:"face_registered"`
	Status         pgtype.Text      `json:"status"`
	CreatedAt      pgtype.Timestamp `json:"created_at"`
	UpdatedAt      pgtype.Timestamp `json:"updated_at"`
}

type User struct {
	ID           pgtype.UUID      `json:"id"`
	Username     string           `json:"username"`
	Email        string           `json:"email"`
	PasswordHash string           `json:"password_hash"`
	FullName     pgtype.Text      `json:"full_name"`
	Phone        pgtype.Text      `json:"phone"`
	RoleID       pgtype.UUID      `json:"role_id"`
	Status       pgtype.Text      `json:"status"`
	LastLoginAt  pgtype.Timestamp `json:"last_login_at"`
	CreatedAt    pgtype.Timestamp `json:"created_at"`
	UpdatedAt    pgtype.Timestamp `json:"updated_at"`
}

type Zone struct {
	ID          pgtype.UUID      `json:"id"`
	Name        string           `json:"name"`
	Description pgtype.Text      `json:"description"`
	CreatedAt   pgtype.Timestamp `json:"created_at"`
}